+++ example/http/client/coro/http_client_coro.cpp
@@ -127,7 +127,7 @@ int main(int argc, char** argv)
     net::io_context ioc;
 
     // Launch the asynchronous operation
-    net::spawn(ioc, std::bind(
+    boost::asio::spawn(ioc, std::bind(
         &do_session,
         std::string(host),
         std::string(port),