+++ example/websocket/client/coro-ssl/websocket_client_coro_ssl.cpp
@@ -153,7 +153,7 @@ int main(int argc, char** argv)
     load_root_certificates(ctx);
 
     // Launch the asynchronous operation
-    net::spawn(ioc, std::bind(
+    boost::asio::spawn(ioc, std::bind(
         &do_session,
         std::string(host),
         std::string(port),