+++ example/websocket/server/coro/websocket_server_coro.cpp
@@ -129,7 +129,7 @@ do_listen(
         if(ec)
             fail(ec, "accept");
         else
-            net::spawn(
+            boost::asio::spawn(
                 acceptor.get_executor(),
                 std::bind(
                     &do_session,
@@ -158,7 +158,7 @@ int main(int argc, char* argv[])
     net::io_context ioc(threads);
 
     // Spawn a listening port
-    net::spawn(ioc,
+    boost::asio::spawn(ioc,
         std::bind(
             &do_listen,
             std::ref(ioc),